+++ taiga/projects/userstories/api.py
@@ -70,6 +70,8 @@ class UserStoryViewSet(OCCResourceMixin, VotedResourceMixin, HistoryResourceMixi
                        base_filters.CreatedDateFilter,
                        base_filters.ModifiedDateFilter,
                        base_filters.FinishDateFilter,
+                       base_filters.MilestoneEstimatedStartFilter,
+                       base_filters.MilestoneEstimatedFinishFilter,
                        base_filters.OrderByFilterMixin)
     filter_fields = ["project",
                      "project__slug",